requirements.txt
streamlit
ezdxf
matplotlib
Pillow